+++ CHANGELOG.md
@@ -7,7 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Added config option for redstone pulse length
+
 ### Changed
+
+- Redstone pulse length for keycard reader and lock block increased
 
 ### Deprecated